add cleanup function
